+++ src/main/java/io/bookwright/di/ApiModule.java
@@ -5,6 +5,7 @@ import com.google.inject.Provides;
 import com.google.inject.Singleton;
 import io.bookwright.api.AuthApi;
 import io.bookwright.api.BookingApi;
+import io.bookwright.api.LocalBookingApi;
 import io.bookwright.api.RetrofitFactory;
 import io.bookwright.config.Configs;
 import io.bookwright.config.MainConfig;
@@ -42,5 +43,11 @@ public class ApiModule extends AbstractModule {
   @Singleton
   BookingApi bookingApi(Retrofit retrofit) {
     return retrofit.create(BookingApi.class);
+  }
+
+  @Provides
+  @Singleton
+  LocalBookingApi localBookingApi(MainConfig config) {
+    return RetrofitFactory.create(config.localBookingBaseUrl()).create(LocalBookingApi.class);
   }
 }